feat(#?): Add analytics scripts
