+++ Assets/Scripts/Analytics/CustomAnalyticsEvent.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 332818b2034ec4d87b965c505cb2bc55
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
